requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
python-dotenv==1.0.1
passlib[argon2,bcrypt]==1.7.4
cryptography==44.0.1
pytest==8.3.3
joblib==1.4.2
